refactor(maru_vllm): move the KV layout geometry into its own module
connector.py is 3.7k lines, and the layout layer added by this PR is a
self-contained ~200-line block of pure module-level code — shapes and
engine-config numbers in, a resolved KVLayout out, no connector state.
The reference implementation draws the same boundary (LMCache keeps this
concern in a dedicated kv_format package), and the known follow-ups
(stride verification, padded pool views) all land in this layer, so give
it its own file before it grows in place.

A pure move: KVLayout, _kv_layout_candidates, _detect_kv_layout,
_canonical_paged_view, _vllm_kv_cache_layout and the worker's
_layout_fits staticmethod (now a module function) relocate verbatim to
maru_vllm/kv_layout.py — AST-identical except the dropped @staticmethod.
The new module imports only torch; the vLLM logger import falls back to
stdlib logging so the pure layout logic stays importable without vLLM.

No behavior change: 154 connector tests and the wider 758 pass unchanged.
